pyproject.toml
[tool.poetry]
name = "sorobonto-backend-api"
version = "0.1.0"
description = ""
authors = ["dawdaborjekujabi@gmail.com <dawdaborjekujabi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
graphene-django = "^3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
